Move get_pe_scaffold_search into data_prep
